+++ src/plugins/nonebot_bison/utils/http.py
@@ -1,0 +1,11 @@
+import functools
+
+import httpx
+
+from ..plugin_config import plugin_config
+
+http_client = functools.partial(
+    httpx.AsyncClient,
+    proxies=plugin_config.bison_proxy or None,
+    headers={"user-agent": plugin_config.bison_ua},
+)