+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 最近更新
+
+### 新功能
+
+- 添加User-Agent配置 [@felinae98](https://github.com/felinae98) ([#78](https://github.com/felinae98/nonebot-bison/pull/78))
+- 增加代理设置 [@felinae98](https://github.com/felinae98) ([#71](https://github.com/felinae98/nonebot-bison/pull/71))
+- 增加Parse Target功能 [@felinae98](https://github.com/felinae98) ([#72](https://github.com/felinae98/nonebot-bison/pull/72))
+
+### Bug 修复
+
+- 捕获SSL异常 [@felinae98](https://github.com/felinae98) ([#75](https://github.com/felinae98/nonebot-bison/pull/75))
+
 ## v0.5.3
 
 - on_command 设置 block=True (#63) @MeetWq